src/requirements-dev.txt
-rrequirements.txt

black==18.9b0
cryptography~=3.1.1
djangorestframework~=3.9.4
flake8==3.7.7
flake8-bugbear==19.3.0
isort==4.3.20
jedi==0.13.3
jsonschema==3.0.1
mkdocs==1.0.4
pem==20.1.0
pre-commit==1.16.1
PyJWT~=1.7.1
pytest-cov==2.7.1
pytest-django==3.5.1
pytest-flake8==1.0.4
seed-isort-config==1.9.1
tox==3.12.1
